requirements.txt
python-dateutil>=2.9.0
requests>=2.32.0
pytest>=8.0.0  # dev / tests
